Created the nedb module handler
In the folder db is located two JS files, the first one is the nedbmodule.js
that is a npm module that handles the Watches and other stuff all related
to the management of the embeded database.

The other file called p.js is a test for the nedbmodule.js.

There is also files.db and folders.db which are the nedb databases.