Added the Electron.io related objects
This commit is part of the first set of commits that includes all the
progress done until 10/02/2016.

Added the index.html which is the main page, the librarymanageer.html
which is the page that handles the watches and library related stuff, the
main.js that loads the whole thing and manages the multi-window feacture.

All the JavaScript, CSS and other sources can be founded in the js, css
and src folders.

Also there is a config.json that stores the last location of the browser
when closed and the main.js use it when it opens.